Added INT typo {def start() -> scoped_session:     engine = create_engine(DB_URI, client_encoding="utf8")     BASE.metadata.bind = engine     BASE.metadata.create_all(engine)     return scoped_session(sessionmaker(bind=engine, autoflush=True))  }

+++ vexana/modules/sql/feds_sql.py
@@ -2,7 +2,7 @@ import threading
 
 from vexana import dispatcher
 from vexana.modules.sql import BASE, SESSION
-from sqlalchemy import Boolean, Column, Integer, String, UnicodeText
+from sqlalchemy import Boolean, Column, BigInteger, String, UnicodeText
 from telegram.error import BadRequest, Unauthorized
 
 
@@ -58,7 +58,7 @@ class BansF(BASE):
 
 class FedsUserSettings(BASE):
     __tablename__ = "feds_settings"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     should_report = Column(Boolean, default=True)
 
     def __init__(self, user_id):
